Combine `jwt` Profile with `secure`
