+++ song-server/src/main/java/bio/overture/song/server/config/SecurityConfig.java
@@ -19,23 +19,34 @@ package bio.overture.song.server.config;
 import bio.overture.song.server.security.EgoApiKeyIntrospector;
 import bio.overture.song.server.security.StudySecurity;
 import bio.overture.song.server.security.SystemSecurity;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Pattern;
 import lombok.Getter;
 import lombok.Setter;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
+import lombok.val;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Profile;
+import org.springframework.http.HttpHeaders;
+import org.springframework.security.authentication.AuthenticationManager;
+import org.springframework.security.authentication.AuthenticationManagerResolver;
+import org.springframework.security.authentication.ProviderManager;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
 import org.springframework.security.config.annotation.web.configuration.WebSecurityConfigurerAdapter;
+import org.springframework.security.oauth2.jwt.JwtDecoder;
+import org.springframework.security.oauth2.server.resource.authentication.JwtAuthenticationProvider;
+import org.springframework.security.oauth2.server.resource.authentication.OpaqueTokenAuthenticationProvider;
 import org.springframework.security.oauth2.server.resource.introspection.OpaqueTokenIntrospector;
 import org.springframework.stereotype.Component;
 import org.springframework.validation.annotation.Validated;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.validation.constraints.NotNull;
+import javax.validation.constraints.Pattern;
+import java.util.UUID;
 
 @Slf4j
 @Getter
@@ -48,84 +59,112 @@ import org.springframework.validation.annotation.Validated;
 @Configuration
 public class SecurityConfig extends WebSecurityConfigurerAdapter {
 
-  @Autowired private SwaggerConfig swaggerConfig;
+    @Autowired
+    private SwaggerConfig swaggerConfig;
+    @Autowired
+    private JwtDecoder jwtDecoder;
 
-  private String introspectionUri;
-  private String clientId;
-  private String clientSecret;
+    private String introspectionUri;
+    private String clientId;
+    private String clientSecret;
 
-  private final ScopeConfig scope = new ScopeConfig();
+    private final ScopeConfig scope = new ScopeConfig();
 
-  @Bean
-  public SystemSecurity systemSecurity() {
-    return new SystemSecurity(scope.getSystem());
-  }
+    @Bean
+    public SystemSecurity systemSecurity() {
+        return new SystemSecurity(scope.getSystem());
+    }
 
-  @Bean
-  public StudySecurity studySecurity(@Autowired SystemSecurity systemSecurity) {
-    return StudySecurity.builder()
-        .studyPrefix(scope.getStudy().getPrefix())
-        .studySuffix(scope.getStudy().getSuffix())
-        .systemScope(scope.getSystem())
-        .build();
-  }
+    @Bean
+    public AuthenticationManagerResolver<HttpServletRequest> tokenAuthenticationManagerResolver() {
 
-  @Bean
-  public OpaqueTokenIntrospector introspector() {
-    return new EgoApiKeyIntrospector(introspectionUri, clientId, clientSecret);
-  }
+        // Auth Managers for JWT and for ApiKeys. JWT uses the default auth provider,
+        // but OpaqueTokens are handled by the custom EgoApiKeyIntrospector
+        AuthenticationManager jwt = new ProviderManager(new JwtAuthenticationProvider(jwtDecoder));
+        AuthenticationManager opaqueToken =
+                new ProviderManager(new OpaqueTokenAuthenticationProvider(new EgoApiKeyIntrospector(introspectionUri, clientId, clientSecret)));
 
-  @Override
-  @SneakyThrows
-  public void configure(HttpSecurity http) {
-    http.authorizeRequests()
-        .antMatchers("/isAlive")
-        .permitAll()
-        .antMatchers("/studies/**")
-        .permitAll()
-        .antMatchers("/upload/**")
-        .permitAll()
-        .antMatchers("/entities/**")
-        .permitAll()
-        .antMatchers("/export/**")
-        .permitAll()
-        .antMatchers("/schemas/**")
-        .permitAll()
-        .antMatchers(swaggerConfig.getAlternateSwaggerUrl())
-        .permitAll()
-        .antMatchers("/swagger**", "/swagger-resources/**", "/v2/api**", "/webjars/**")
-        .permitAll()
-        .and()
-        .authorizeRequests()
-        .anyRequest()
-        .authenticated();
+        return (request) -> useJwt(request) ? jwt : opaqueToken;
+    }
 
-    http.oauth2ResourceServer(
-        oauth2 ->
-            oauth2.opaqueToken(
-                opaqueToken ->
-                    opaqueToken.introspector(
-                        new EgoApiKeyIntrospector(introspectionUri, clientId, clientSecret))));
-  }
+    @Bean
+    public StudySecurity studySecurity(@Autowired SystemSecurity systemSecurity) {
+        return StudySecurity.builder()
+                .studyPrefix(scope.getStudy().getPrefix())
+                .studySuffix(scope.getStudy().getSuffix())
+                .systemScope(scope.getSystem())
+                .build();
+    }
 
-  @Getter
-  @Setter
-  public static class ScopeConfig {
+    @Bean
+    public OpaqueTokenIntrospector introspector() {
+        return new EgoApiKeyIntrospector(introspectionUri, clientId, clientSecret);
+    }
 
-    @NotNull private String system;
-    private final StudyScopeConfig study = new StudyScopeConfig();
+    @Override
+    @SneakyThrows
+    public void configure(HttpSecurity http) {
+        http.authorizeRequests()
+                .antMatchers("/isAlive")
+                .permitAll()
+                .antMatchers("/studies/**")
+                .permitAll()
+                .antMatchers("/upload/**")
+                .permitAll()
+                .antMatchers("/entities/**")
+                .permitAll()
+                .antMatchers("/export/**")
+                .permitAll()
+                .antMatchers("/schemas/**")
+                .permitAll()
+                .antMatchers(swaggerConfig.getAlternateSwaggerUrl())
+                .permitAll()
+                .antMatchers("/swagger**", "/swagger-resources/**", "/v2/api**", "/webjars/**")
+                .permitAll()
+                .and()
+                .authorizeRequests()
+                .anyRequest()
+                .authenticated();
+
+        http.oauth2ResourceServer(
+                oauth2 -> oauth2.authenticationManagerResolver(this.tokenAuthenticationManagerResolver()));
+    }
 
     @Getter
     @Setter
-    public static class StudyScopeConfig {
+    public static class ScopeConfig {
 
-      @NotNull
-      @Pattern(regexp = "^\\w+\\W$")
-      private String prefix;
+        @NotNull
+        private String system;
+        private final StudyScopeConfig study = new StudyScopeConfig();
 
-      @NotNull
-      @Pattern(regexp = "^\\W\\w+$")
-      private String suffix;
+        @Getter
+        @Setter
+        public static class StudyScopeConfig {
+
+            @NotNull
+            @Pattern(regexp = "^\\w+\\W$")
+            private String prefix;
+
+            @NotNull
+            @Pattern(regexp = "^\\W\\w+$")
+            private String suffix;
+        }
     }
-  }
+
+    private boolean useJwt(HttpServletRequest request) {
+        val authorizationHeaderValue = request.getHeader(HttpHeaders.AUTHORIZATION);
+        if (authorizationHeaderValue != null && authorizationHeaderValue.startsWith("Bearer")) {
+            String token = authorizationHeaderValue.substring(7);
+            try {
+                UUID.fromString(token);
+                // able to parse as UUID, so this token matches our EgoApiKey format
+                return false;
+            } catch (IllegalArgumentException e) {
+                // unable to parse as UUID, use our JWT resolvers
+                return true;
+            }
+        }
+        return true;
+    }
 }
